server/requirements.txt
Django==2.2.16
djangorestframework==3.11.1
requests
psutil